Moving classes to storage package
