+++ app/src/main/kotlin/com/arturogutierrez/openticator/di/component/ApplicationComponent.kt
@@ -4,8 +4,8 @@ import android.content.Context
 import android.content.SharedPreferences
 import android.view.LayoutInflater
 import com.arturogutierrez.openticator.di.module.ApplicationModule
-import com.arturogutierrez.openticator.domain.DatabaseConfigurator
-import com.arturogutierrez.openticator.domain.Preferences
+import com.arturogutierrez.openticator.storage.database.DatabaseConfigurator
+import com.arturogutierrez.openticator.storage.preferences.Preferences
 import com.arturogutierrez.openticator.domain.account.repository.AccountRepository
 import com.arturogutierrez.openticator.domain.category.CategorySelector
 import com.arturogutierrez.openticator.domain.category.repository.CategoryRepository
