+++ app/src/main/kotlin/com/arturogutierrez/openticator/di/module/ApplicationModule.kt
@@ -3,9 +3,9 @@ package com.arturogutierrez.openticator.di.module
 import android.content.Context
 import android.content.SharedPreferences
 import android.view.LayoutInflater
-import com.arturogutierrez.openticator.domain.DatabaseConfigurator
-import com.arturogutierrez.openticator.domain.Preferences
-import com.arturogutierrez.openticator.domain.PreferencesImpl
+import com.arturogutierrez.openticator.storage.database.DatabaseConfigurator
+import com.arturogutierrez.openticator.storage.preferences.Preferences
+import com.arturogutierrez.openticator.storage.preferences.PreferencesImpl
 import com.arturogutierrez.openticator.domain.account.repository.AccountRepository
 import com.arturogutierrez.openticator.domain.account.repository.AccountRepositoryImpl
 import com.arturogutierrez.openticator.domain.category.CategorySelector
@@ -17,9 +17,9 @@ import com.arturogutierrez.openticator.executor.JobExecutor
 import com.arturogutierrez.openticator.executor.MainThread
 import com.arturogutierrez.openticator.executor.PostExecutionThread
 import com.arturogutierrez.openticator.executor.ThreadExecutor
-import com.arturogutierrez.openticator.storage.AccountDiskDataStore
-import com.arturogutierrez.openticator.storage.CategoryDiskDataStore
-import com.arturogutierrez.openticator.storage.RealmDatabaseConfigurator
+import com.arturogutierrez.openticator.storage.database.AccountDiskDataStore
+import com.arturogutierrez.openticator.storage.database.CategoryDiskDataStore
+import com.arturogutierrez.openticator.storage.database.RealmDatabaseConfigurator
 import dagger.Module
 import dagger.Provides
 import javax.inject.Singleton
