+++ data/src/main/kotlin/com/arturogutierrez/openticator/storage/database/CategoryDiskDataStore.kt
@@ -1,4 +1,4 @@
-package com.arturogutierrez.openticator.storage
+package com.arturogutierrez.openticator.storage.database
 
 import com.arturogutierrez.openticator.domain.account.model.Account
 import com.arturogutierrez.openticator.domain.category.model.Category
