+++ data/src/main/kotlin/com/arturogutierrez/openticator/storage/preferences/PreferencesImpl.kt
@@ -1,6 +1,7 @@
-package com.arturogutierrez.openticator.domain
+package com.arturogutierrez.openticator.storage.preferences
 
 import android.content.SharedPreferences
+import com.arturogutierrez.openticator.storage.preferences.Preferences
 import javax.inject.Inject
 
 class PreferencesImpl @Inject constructor(val sharedPreferences: SharedPreferences) : Preferences {
