+++ domain/src/main/kotlin/com/arturogutierrez/openticator/domain/bootstrap/InitialScreenSelector.kt
@@ -1,6 +1,6 @@
 package com.arturogutierrez.openticator.domain.bootstrap
 
-import com.arturogutierrez.openticator.domain.Preferences
+import com.arturogutierrez.openticator.storage.preferences.Preferences
 import javax.inject.Inject
 
 class InitialScreenSelector @Inject constructor(private val preferences: Preferences) {
