+++ domain/src/main/kotlin/com/arturogutierrez/openticator/domain/bootstrap/StorageInitializator.kt
@@ -1,7 +1,7 @@
 package com.arturogutierrez.openticator.domain.bootstrap
 
-import com.arturogutierrez.openticator.domain.DatabaseConfigurator
-import com.arturogutierrez.openticator.domain.Preferences
+import com.arturogutierrez.openticator.storage.database.DatabaseConfigurator
+import com.arturogutierrez.openticator.storage.preferences.Preferences
 import javax.inject.Inject
 
 class StorageInitializator @Inject constructor(val preferences: Preferences, val databaseConfigurator: DatabaseConfigurator) {
