+++ domain/src/main/kotlin/com/arturogutierrez/openticator/storage/database/DatabaseConfigurator.kt
@@ -1,4 +1,4 @@
-package com.arturogutierrez.openticator.domain
+package com.arturogutierrez.openticator.storage.database
 
 interface DatabaseConfigurator {
   fun configure(passwordInBase64: String)
