+++ domain/src/main/kotlin/com/arturogutierrez/openticator/storage/preferences/Preferences.kt
@@ -1,4 +1,4 @@
-package com.arturogutierrez.openticator.domain
+package com.arturogutierrez.openticator.storage.preferences
 
 interface Preferences {
   companion object {
